requirements.txt
pyyaml
opencv-python
numpy
pytesseract
pywin32
jinja2
keyboard